requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
Telethon==1.25.1
urllib3==1.26.12
